Consolidating MagicLinkSignInFragment with SignInFragment
